+++ src/backend/mailbox_manager/api/client/viewsets.py
@@ -33,6 +33,7 @@ class MailDomainViewSet(
 
     POST /api/<version>/mail-domains/ with expected data:
         - name: str
+        - support_email: str
         Return newly created domain
     """