+++ src/backend/mailbox_manager/tests/api/mail_domain/test_api_mail_domains_retrieve.py
@@ -87,4 +87,5 @@ def test_api_mail_domains__retrieve_authenticated_related():
         "updated_at": domain.updated_at.isoformat().replace("+00:00", "Z"),
         "abilities": domain.get_abilities(user),
         "count_mailboxes": 10,
+        "support_email": domain.support_email,
     }